+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- `microphone` command
 - `webcamgif` command
 - Notification on command execution (except for owner)
 - Documentation for configuration and account creation